Adds language switcher component
Adds a language switcher component to the navbar.

The language switcher allows users to select their preferred language from a dropdown menu.
It includes the flags of Mexico and the USA as visual cues for the Spanish and English languages, respectively.

This enhances the user experience by enabling easy switching between language settings.